backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.7.0
pdfplumber==0.11.0
python-multipart==0.0.9